requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
apitally==0.11.0
asgiref==3.8.1
assemblyai==0.28.1
backoff==2.2.1
beautifulsoup4==4.12.3
billiard==4.2.0
boto3==1.34.136
botocore==1.34.136
cachetools==5.4.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
defusedxml==0.7.1
distro==1.9.0
Django==5.0.6
django-environ==0.9.0
djangorestframework==3.15.2
exceptiongroup==1.2.0
fastavro==1.9.5
fastembed==0.3.4
feedparser==6.0.11
ffmpeg-python==0.2.0
filelock==3.15.4
flatbuffers==24.3.25
flower==2.0.1
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
humanfriendly==10.0
humanize==4.9.0
idna==3.7
inflection==0.5.1
jiter==0.5.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.3.7
langsmith==0.1.93
loguru==0.7.2
mmh3==4.1.0
mpmath==1.3.0
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.18.1
openai==1.42.0
orjson==3.10.5
packaging==24.1
parameterized==0.9.0
pgvector==0.3.2
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pyloudnorm==0.1.1
PyStemmer==2.2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.3
redis==5.0.7
regex==2024.5.15
requests==2.32.3
resend==2.1.0
rsa==4.9
s3transfer==0.10.2
scipy==1.14.0
sentry-sdk==2.10.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soundfile==0.12.1
soupsieve==2.5
sqlparse==0.5.0
sseclient-py==1.8.0
sympy==1.13.0
tenacity==8.4.2
tiktoken==0.7.0
tokenizers==0.19.1
tornado==6.4.1
tqdm==4.66.4
types-requests==2.32.0.20240622
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
whitenoise==6.7.0
